Revert "Revert "Merge branch 'master' of https://github.com/simon2000234/MyTunes""
This reverts commit 00e65a193a00d6685da738fcb04a3ae161ec3a84.

+++ src/mytunes/gui/FXMLDocumentController.java
@@ -18,6 +18,9 @@ import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
+import javafx.scene.control.Alert;
+import javafx.scene.control.Alert.AlertType;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.Label;
 import javafx.scene.control.ListView;
 import javafx.scene.control.Slider;
@@ -112,8 +115,12 @@ public class FXMLDocumentController implements Initializable
                 model.setSelectedSong(sopview.getSelectionModel().getSelectedItem());
             }
 
+
         }
         );
+
+        });
+
 
     }
 
@@ -179,6 +186,7 @@ public class FXMLDocumentController implements Initializable
     @FXML
     private void handleplaylistdelete(ActionEvent event)
     {
+
         Playlist pl = plview.getSelectionModel().getSelectedItem();
         try
         {
@@ -186,6 +194,25 @@ public class FXMLDocumentController implements Initializable
         } catch (SQLException ex)
         {
             Logger.getLogger(FXMLDocumentController.class.getName()).log(Level.SEVERE, null, ex);
+
+        Alert confirmPLDelete = new Alert(AlertType.CONFIRMATION, "Delete: "
+                + model.getSelectedPlaylist().getPlaylistName() + "?",
+                ButtonType.YES, ButtonType.NO, ButtonType.CANCEL);
+        confirmPLDelete.setTitle("Delete playlist");
+        confirmPLDelete.setHeaderText("Are you sure?");
+        confirmPLDelete.showAndWait();
+
+        Playlist pl = plview.getSelectionModel().getSelectedItem();
+        if (confirmPLDelete.getResult() == ButtonType.YES)
+        {        
+            try
+            {
+                model.deleteplaylist(pl);
+            } catch (SQLException ex)
+            {
+                Logger.getLogger(FXMLDocumentController.class.getName()).log(Level.SEVERE, null, ex);
+            }
+
         }
     }
 
@@ -202,13 +229,24 @@ public class FXMLDocumentController implements Initializable
     @FXML
     private void handlesopdelete(ActionEvent event)
     {
-        model.removeSong(model.getSelectedPlaylist(), model.getSelectedSong());
-        try
+        Alert confirmSOPDelete = new Alert(AlertType.CONFIRMATION, "Delete: "
+                + model.getSelectedSong().getTitle() + "\n"
+                + "From: " + model.getSelectedPlaylist().getPlaylistName() + "?",
+                ButtonType.YES, ButtonType.NO, ButtonType.CANCEL);
+        confirmSOPDelete.setHeaderText("Are you sure?");
+        confirmSOPDelete.setTitle("Delete song from playlist");
+        confirmSOPDelete.showAndWait();
+        if (confirmSOPDelete.getResult() == ButtonType.YES)
         {
-            sopview.setItems(model.updateSopview());
-        } catch (SQLException ex)
-        {
-            Logger.getLogger(FXMLDocumentController.class.getName()).log(Level.SEVERE, null, ex);
+
+            model.removeSong(model.getSelectedPlaylist(), model.getSelectedSong());
+            try
+            {
+                sopview.setItems(model.updateSopview());
+            } catch (SQLException ex)
+            {
+                Logger.getLogger(FXMLDocumentController.class.getName()).log(Level.SEVERE, null, ex);
+            }
         }
     }
 
@@ -221,9 +259,17 @@ public class FXMLDocumentController implements Initializable
     @FXML
     private void handlesongsdelete(ActionEvent event)
     {
-        Song ds = songsview.getSelectionModel().getSelectedItem();
-        model.deleteSong(ds);
+        Alert confirmDelete = new Alert(AlertType.CONFIRMATION, "Delete: " + model.getSelectedSong().getTitle() + "?", ButtonType.YES, ButtonType.NO, ButtonType.CANCEL);
+        confirmDelete.setTitle("Delete Song");
+        confirmDelete.setHeaderText("Are you sure?");
+        confirmDelete.showAndWait();
 
+        if (confirmDelete.getResult() == ButtonType.YES)
+        {
+            Song ds = songsview.getSelectionModel().getSelectedItem();
+            model.deleteSong(ds);
+
+        }
     }
 
     @FXML
@@ -277,10 +323,10 @@ public class FXMLDocumentController implements Initializable
                 || model.getCurPlaySong() != model.getSelectedSong().getFilePath())
         {
             model.StopSong();
-//            model.setSelectedSong(songsview.getSelectionModel().getSelectedItem());
             model.PlaySong();
             isPaused = true;
             headlinelbl.setText("Currently playing: " + model.getSelectedSong().getTitle());
+            model.playNextSong(new Song(1, "Booby Drake", "some dude", 163, "sdas", "Data/BenJamin_Banger_-_01_-_Bobby_Drake.mp3"));
         } else
         {
             model.PausePlaySong();
